requirements.txt
streamlit==1.33.0
Pillow==10.4.0
requests==2.31.0
google-generativeai==0.4.1